pyproject.toml
[project]
name = "lang-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipywidgets>=8.1.8",
    "jsonpatch>=1.33",
    "jsonpointer>=2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=1.0.1",
    "tavily-python>=0.7.23",
]
